requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.6.2.post1
    # via
    #   -r requirements.in
    #   starlette
    #   watchfiles
build==1.2.2.post1
    # via -r requirements.in
click==8.1.7
    # via
    #   -r requirements.in
    #   uvicorn
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
contourpy==1.3.1
    # via
    #   -r requirements.in
    #   matplotlib
cycler==0.12.1
    # via
    #   -r requirements.in
    #   matplotlib
et-xmlfile==2.0.0
    # via
    #   -r requirements.in
    #   openpyxl
fastapi==0.115.5
    # via -r requirements.in
fonttools==4.55.0
    # via
    #   -r requirements.in
    #   matplotlib
h11==0.14.0
    # via
    #   -r requirements.in
    #   uvicorn
httptools==0.6.4
    # via -r requirements.in
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
jinja2==3.1.4
    # via -r requirements.in
kiwisolver==1.4.7
    # via
    #   -r requirements.in
    #   matplotlib
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
matplotlib==3.9.2
    # via -r requirements.in
numpy==2.1.3
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
openpyxl==3.1.5
    # via -r requirements.in
packaging==24.2
    # via
    #   -r requirements.in
    #   build
    #   matplotlib
pandas==2.2.3
    # via -r requirements.in
pillow==11.0.0
    # via
    #   -r requirements.in
    #   matplotlib
pydantic==2.10.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.1
    # via
    #   -r requirements.in
    #   pydantic
pyparsing==3.2.0
    # via
    #   -r requirements.in
    #   matplotlib
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.18
    # via -r requirements.in
pytz==2024.2
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
starlette==0.41.3
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2024.2
    # via
    #   -r requirements.in
    #   pandas
uvicorn==0.32.1
    # via -r requirements.in
watchfiles==1.0.0
    # via -r requirements.in
websockets==14.1
    # via -r requirements.in
wheel==0.45.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
